+++ eslint.config.mjs
@@ -30,6 +30,7 @@ export default [
       globals: {
         ...globals.browser,
         ...globals.jest,
+        ...globals.vitest,
       },
     },
   },